Fixed Calendar Issues
